Data Science/Web_Scraper/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
get==2019.4.13
idna==2.9
joblib==0.15.1
nltk==3.5
numpy==1.18.5
pandas==1.0.4
pipe==1.5.0
post==2019.4.13
public==2019.4.13
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
query-string==2019.4.13
regex==2020.6.8
request==2019.4.13
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
sspipe==0.1.14
tqdm==4.46.1
urllib3==1.25.9
